Cloud-Automation/Azure/LAB01-VM-Creation-With-AzureSDK/requirements.txt
azure-identity>=1.12.0
azure-mgmt-resource>=22.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0 